Lesson 089

+++ MyCourse/Models/Options/CoursesOptions.cs
@@ -8,7 +8,7 @@ namespace MyCourse.Models.Options
 {
     public class CoursesOptions
     {
-        public long PerPage { get; set; }
+        public int PerPage { get; set; }
         public long CacheDuration { get; set; }
         public CoursesOrderOptions Order { get; set; }
     }
